Fix alloc_realloc UB
 void ** does not exist: the address of a generic pointer is not
properly defined (different pointer types may have different
representations). So, alloc_realloc cannot exist as is without UB.
Fortunately, it's not supposed to be used in the skalibs programming
style, and skalibs itself only uses it in two places
(stralloc_ready_tuned and stralloc_shrink) where the pointer is a
char *.
 So we just fix the UB by making alloc_realloc() take a char **,
and it's only defined for that pointer type.
 Nothing to see here folks, nothing happened at all.